+++ client/index.html
@@ -1,49 +1,55 @@
 <!DOCTYPE html>
 <html lang="en">
+
 <head>
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>SPA Fancy Todo</title>
     <style>
-        body{
+        body {
             font-family: 'Courier New', Courier, monospace;
         }
-        table{
+
+        table {
             border-collapse: collapse;
         }
     </style>
 </head>
+
 <body>
-    <h1>Hallo</h1>
+    <h1>Fancy Todo App</h1>
+    <h2 id="welcoming"></h2>
     <form action="" id="form-login">
-        <div>
-            <label for="username">Username</label><br>
-            <input type="text" name="username" id="username">
-        </div>
         <div>
             <label for="email">Email</label><br>
             <input type="text" name="email" id="email">
-        </div><div>
-            <label for="password">password</label><br>
+        </div>
+        <div>
+            <label for="password">Password</label><br>
             <input type="password" name="password" id="password">
         </div>
+        <br>
         <button type="submit">Click</button>
     </form>
+
     <div id="home">
         <button id="logout">Logout</button>
     </div>
-    <div>
-        <p>Data</p>
-        <table border="1" id="findAll">
-            <tr>
-                <th>Title</th>
-                <th>Description</th>
-                <th>Status</th>
-                <th>Due Date</th>
-            </tr>
-        </table>
-    </div>
+    <br>
+
+
+
+    <table border="1" id="todoList">
+        <tr>
+            <th>Title</th>
+            <th>Description</th>
+            <th>Status</th>
+            <th>Due Date</th>
+        </tr>
+    </table>
+
     <script src="https://code.jquery.com/jquery-3.5.1.min.js" integrity="sha256-9/aliU8dGd2tb6OSsuzixeV4y/faTqgFtohetphbbj0=" crossorigin="anonymous"></script>
-    <script src="./helper/helper.js"></script>
+    <script src="./main.js"></script>
 </body>
+
 </html>
